+++ src/interfaces/queue-options.ts
@@ -63,6 +63,9 @@ export interface QueueOptions extends QueueBaseOptions {
  * Options for QueueEvents
  */
 export interface QueueEventsOptions extends QueueBaseOptions {
+  /**
+   * Condition to start listening to events at instance creation.
+   */
   autorun?: boolean;
   /**
    * Last event Id. If provided it is possible to continue