+++ src/interfaces/worker-options.ts
@@ -10,6 +10,9 @@ export type Processor<T = any, R = any, N extends string = string> = (
 ) => Promise<R>;
 
 export interface WorkerOptions extends QueueBaseOptions {
+  /**
+   * Condition to start processor at instance creation.
+   */
   autorun?: boolean;
   /**
    * Amount of jobs that a single worker is allowed to work on