+++ src/test/test_obliterate.ts
@@ -1,11 +1,12 @@
-import { Queue, QueueEvents, Worker } from '../classes';
-import { delay, removeAllQueueData } from '../utils';
 import { expect } from 'chai';
 import * as IORedis from 'ioredis';
+import { after } from 'lodash';
 import { beforeEach, describe, it } from 'mocha';
 import { v4 } from 'uuid';
+import { Queue, QueueEvents, FlowProducer, Worker } from '../classes';
+import { delay, removeAllQueueData } from '../utils';
 
-describe('Obliterate', () => {
+describe('Obliterate', function() {
   let queue: Queue;
   let queueEvents: QueueEvents;
   let queueName: string;
@@ -17,7 +18,7 @@ describe('Obliterate', () => {
     await queueEvents.waitUntilReady();
   });
 
-  afterEach(async () => {
+  afterEach(async function() {
     await queue.close();
     await queueEvents.close();
     await removeAllQueueData(new IORedis(), queueName);
@@ -60,6 +61,54 @@ describe('Obliterate', () => {
     expect(keys.length).to.be.eql(0);
 
     await worker.close();
+  });
+
+  describe('when creating a flow', async () => {
+    it('obliterates a queue with jobs and its dependency keys', async () => {
+      await queue.waitUntilReady();
+      const name = 'child-job';
+
+      let first = true;
+      const worker = new Worker(queue.name, async job => {
+        if (first) {
+          first = false;
+          throw new Error('failed first');
+        }
+        return delay(10);
+      });
+      await worker.waitUntilReady();
+
+      const completing = new Promise((resolve, reject) => {
+        worker.on('completed', after(2, resolve));
+      });
+
+      const failing = new Promise((resolve, reject) => {
+        worker.on('failed', resolve);
+      });
+
+      const flow = new FlowProducer();
+      await flow.add({
+        name: 'parent-job',
+        queueName: queueName,
+        data: {},
+        children: [
+          { name, data: { idx: 0, foo: 'bar' }, queueName },
+          { name, data: { idx: 1, foo: 'baz' }, queueName },
+          { name, data: { idx: 2, foo: 'qux' }, queueName },
+        ],
+      });
+
+      await failing;
+      await completing;
+      await queue.obliterate();
+
+      const client = await queue.client;
+      const keys = await client.keys(`bull:${queue.name}:*`);
+
+      expect(keys.length).to.be.eql(0);
+
+      await worker.close();
+    });
   });
 
   it('should raise exception if queue has active jobs', async () => {
@@ -161,7 +210,8 @@ describe('Obliterate', () => {
     expect(logs).to.have.length(0);
   });
 
-  it('should obliterate a queue with high number of jobs in different statuses', async () => {
+  it('should obliterate a queue with high number of jobs in different statuses', async function() {
+    this.timeout(6000);
     const arr1 = [];
     for (let i = 0; i < 300; i++) {
       arr1.push(queue.add('test', { foo: `barLoop${i}` }));
@@ -188,12 +238,9 @@ describe('Obliterate', () => {
 
     const [lastFailedJob] = (await Promise.all(arr2)).splice(-1);
 
-    try {
-      await lastFailedJob.waitUntilFinished(queueEvents);
-      expect(true).to.be.equal(false);
-    } catch (err) {
-      expect(true).to.be.equal(true);
-    }
+    await expect(
+      lastFailedJob.waitUntilFinished(queueEvents),
+    ).to.be.eventually.rejectedWith('failed job');
 
     const arr3 = [];
     for (let i = 0; i < 1623; i++) {
@@ -205,5 +252,7 @@ describe('Obliterate', () => {
     const client = await queue.client;
     const keys = await client.keys(`bull:${queue.name}*`);
     expect(keys.length).to.be.eql(0);
-  }).timeout(20000);
+
+    await worker.close();
+  });
 });